+++ .playground/playgroud.tsx
@@ -1,75 +1,31 @@
 import React, { Fragment } from 'react';
-import {
-  Axis,
-  Chart,
-  getAxisId,
-  getSpecId,
-  Position,
-  ScaleType,
-  Settings,
-  BarSeries,
-  LineSeries,
-  AreaSeries,
-} from '../src';
+import { Axis, Chart, getAxisId, getSpecId, Position, ScaleType, BarSeries } from '../src';
 
 export class Playground extends React.Component {
   render() {
+    const data = [{ x: 0, y: -4 }, { x: 1, y: -3 }, { x: 2, y: 2 }, { x: 3, y: 1 }];
     return (
       <Fragment>
         <div className="chart">
-          <Chart className="story-chart">
-            <Settings
-              theme={{
-                areaSeriesStyle: {
-                  point: {
-                    visible: true,
-                  },
-                },
-              }}
-              xDomain={{
-                max: 3.8,
-              }}
-            />
+          <Chart>
+            <Axis id={getAxisId('top')} position={Position.Bottom} title={'Top axis'} />
             <Axis
-              id={getAxisId('bottom')}
-              position={Position.Bottom}
-              title={'Bottom axis'}
-              showOverlappingTicks={true}
-            />
-            <Axis
-              id={getAxisId('left')}
+              id={getAxisId('left2')}
               title={'Left axis'}
               position={Position.Left}
-              domain={{
-                max: 5,
-              }}
+              tickFormat={(d: any) => Number(d).toFixed(2)}
             />
 
             <BarSeries
-              id={getSpecId('bar')}
+              id={getSpecId('bars')}
               xScaleType={ScaleType.Linear}
               yScaleType={ScaleType.Linear}
-              xAccessor={0}
-              yAccessors={[1]}
-              data={[[0, 1], [1, 2], [2, 10], [3, 4], [4, 5]]}
-            />
-
-            <LineSeries
-              id={getSpecId('line')}
-              xScaleType={ScaleType.Linear}
-              yScaleType={ScaleType.Linear}
-              xAccessor={0}
-              yAccessors={[1]}
-              data={[[0, 1], [1, 2], [2, 10], [3, 4], [4, 5]]}
-            />
-
-            <AreaSeries
-              id={getSpecId('area')}
-              xScaleType={ScaleType.Linear}
-              yScaleType={ScaleType.Linear}
-              xAccessor={0}
-              yAccessors={[1]}
-              data={[[0, 1], [1, 2], [2, 10], [3, 4], [4, 5]]}
+              xAccessor="x"
+              yAccessors={['y']}
+              splitSeriesAccessors={['g']}
+              stackAccessors={['x']}
+              data={data}
+              yScaleToDataExtent={true}
             />
           </Chart>
         </div>