+++ src/chart_types/xy_chart/rendering/rendering.test.ts
@@ -108,8 +108,14 @@ describe('Rendering utils', () => {
       isSeriesVisible: true,
       isLegendItemVisible: true,
       displayValue: {
-        raw: '',
-        formatted: '',
+        formatted: {
+          y0: null,
+          y1: null,
+        },
+        raw: {
+          y0: null,
+          y1: null,
+        },
       },
     };