+++ src/chart_types/xy_chart/utils/series.test.ts
@@ -104,7 +104,8 @@ describe('Series', () => {
         data: [{ x: 1, y1: 21 }, { x: 3, y1: 23 }],
       },
     ];
-    const stackedValues = formatStackedDataSeriesValues(dataSeries, false);
+    const xValues = new Set([1, 2, 3, 4]);
+    const stackedValues = formatStackedDataSeriesValues(dataSeries, false, false, xValues, ScaleType.Linear);
     expect(stackedValues).toMatchSnapshot();
   });
   test('Can stack multiple dataseries', () => {
@@ -142,7 +143,8 @@ describe('Series', () => {
         data: [{ x: 1, y1: 1 }, { x: 2, y1: 2 }, { x: 3, y1: 3 }, { x: 4, y1: 4 }],
       },
     ];
-    const stackedValues = formatStackedDataSeriesValues(dataSeries, false);
+    const xValues = new Set([1, 2, 3, 4]);
+    const stackedValues = formatStackedDataSeriesValues(dataSeries, false, false, xValues, ScaleType.Linear);
     expect(stackedValues).toMatchSnapshot();
   });
   test('Can stack unsorted dataseries', () => {
@@ -164,7 +166,8 @@ describe('Series', () => {
         data: [{ x: 3, y1: 23 }, { x: 1, y1: 21 }],
       },
     ];
-    const stackedValues = formatStackedDataSeriesValues(dataSeries, false);
+    const xValues = new Set([1, 2, 3, 4]);
+    const stackedValues = formatStackedDataSeriesValues(dataSeries, false, false, xValues, ScaleType.Linear);
     expect(stackedValues).toMatchSnapshot();
   });
   test('Can stack high volume of dataseries', () => {
@@ -187,7 +190,8 @@ describe('Series', () => {
         data: new Array(maxArrayItems).fill(0).map((d, i) => ({ x: i, y1: i })),
       },
     ];
-    const stackedValues = formatStackedDataSeriesValues(dataSeries, false);
+    const xValues = new Set(new Array(maxArrayItems).fill(0).map((d, i) => i));
+    const stackedValues = formatStackedDataSeriesValues(dataSeries, false, false, xValues, ScaleType.Linear);
     expect(stackedValues).toMatchSnapshot();
   });
   test('Can stack simple dataseries with scale to extent', () => {
@@ -209,7 +213,8 @@ describe('Series', () => {
         data: [{ x: 1, y1: 21 }, { x: 3, y1: 23 }],
       },
     ];
-    const stackedValues = formatStackedDataSeriesValues(dataSeries, true);
+    const xValues = new Set([1, 2, 3, 4]);
+    const stackedValues = formatStackedDataSeriesValues(dataSeries, true, false, xValues, ScaleType.Linear);
     // the datum on the snapshots is undefined because we are not adding it to
     // the test raw dataseries
     expect(stackedValues).toMatchSnapshot();
@@ -249,7 +254,8 @@ describe('Series', () => {
         data: [{ x: 1, y1: 1 }, { x: 2, y1: 2 }, { x: 3, y1: 3 }, { x: 4, y1: 4 }],
       },
     ];
-    const stackedValues = formatStackedDataSeriesValues(dataSeries, true);
+    const xValues = new Set([1, 2, 3, 4]);
+    const stackedValues = formatStackedDataSeriesValues(dataSeries, true, false, xValues, ScaleType.Linear);
     // the datum on the snapshots is undefined because we are not adding it to
     // the test raw dataseries
     expect(stackedValues).toMatchSnapshot();
@@ -273,7 +279,8 @@ describe('Series', () => {
         data: [{ x: 1, y1: 2, y0: 1 }, { x: 2, y1: 3, y0: 1 }, { x: 3, y1: 23, y0: 4 }, { x: 4, y1: 4, y0: 1 }],
       },
     ];
-    const stackedValues = formatStackedDataSeriesValues(dataSeries, true);
+    const xValues = new Set([1, 2, 3, 4]);
+    const stackedValues = formatStackedDataSeriesValues(dataSeries, true, false, xValues, ScaleType.Linear);
     // the datum on the snapshots is undefined because we are not adding it to
     // the test raw dataseries
 
@@ -308,7 +315,8 @@ describe('Series', () => {
         data: [{ x: 1, y1: 2, y0: 1 }, { x: 2, y1: 3, y0: 1 }, { x: 3, y1: 23, y0: 4 }, { x: 4, y1: 4, y0: 1 }],
       },
     ];
-    const stackedValues = formatStackedDataSeriesValues(dataSeries, true);
+    const xValues = new Set([1, 2, 3, 4]);
+    const stackedValues = formatStackedDataSeriesValues(dataSeries, true, false, xValues, ScaleType.Linear);
     // the datum on the snapshots is undefined because we are not adding it to
     // the test raw dataseries
     expect(stackedValues[0].data[0].y0).toBe(1);
@@ -383,10 +391,16 @@ describe('Series', () => {
       data: TestDataset.BARCHART_2Y0G,
       hideInLegend: false,
     };
+    const xValues = new Set([0, 1, 2, 3]);
     seriesSpecs.set(spec1.id, spec1);
     seriesSpecs.set(spec2.id, spec2);
     const splittedDataSeries = getSplittedSeries(seriesSpecs);
-    const stackedDataSeries = getFormattedDataseries([spec1, spec2], splittedDataSeries.splittedSeries);
+    const stackedDataSeries = getFormattedDataseries(
+      [spec1, spec2],
+      splittedDataSeries.splittedSeries,
+      xValues,
+      ScaleType.Linear,
+    );
     expect(stackedDataSeries.stacked).toMatchSnapshot();
   });
   test('should get series color map', () => {