+++ src/components/legend/legend.tsx
@@ -7,6 +7,8 @@ import { ChartStore } from '../../chart_types/xy_chart/store/chart_state';
 import { Position } from '../../chart_types/xy_chart/utils/specs';
 import { LegendItem } from './legend_item';
 import { Theme } from '../../utils/themes/theme';
+import { TooltipLegendValue } from '../../chart_types/xy_chart/tooltip/tooltip';
+import { BandedAccessorType } from '../../chart_types/xy_chart/rendering/rendering';
 
 interface LegendProps {
   chartStore?: ChartStore; // FIX until we find a better way on ts mobx
@@ -136,29 +138,52 @@ class LegendComponent extends React.Component<LegendProps, LegendState> {
     this.props.chartStore!.onLegendItemOut();
   };
 
-  private renderLegendElement = (item: SeriesLegendItem) => {
-    const { key, displayValue } = item;
-    const { legendPosition, legendItemTooltipValues, isCursorOnChart } = this.props.chartStore!;
-    const tooltipValues = legendItemTooltipValues.get();
-    let tooltipValue;
-
-    if (tooltipValues && tooltipValues.get(key)) {
-      tooltipValue = tooltipValues.get(key);
+  private getLegendValues(
+    tooltipValues: Map<string, TooltipLegendValue> | undefined,
+    key: string,
+    banded: boolean = false,
+  ): any[] {
+    const values = tooltipValues && tooltipValues.get(key);
+    if (values === null || values === undefined) {
+      return banded ? ['', ''] : [''];
     }
 
-    const newDisplayValue = tooltipValue != null ? tooltipValue : '';
+    const { y0, y1 } = values;
+    return banded ? [y1, y0] : [y1];
+  }
 
-    return (
-      <LegendItem
-        {...item}
-        key={key}
-        legendItemKey={key}
-        legendPosition={legendPosition.get()}
-        displayValue={isCursorOnChart.get() ? newDisplayValue : displayValue.formatted}
-        onMouseEnter={this.onLegendItemMouseover(key)}
-        onMouseLeave={this.onLegendItemMouseout}
-      />
-    );
+  private getItemLabel(
+    { banded, label, y1AccessorFormat, y0AccessorFormat }: SeriesLegendItem,
+    yAccessor: BandedAccessorType,
+  ) {
+    if (!banded) {
+      return label;
+    }
+
+    return yAccessor === BandedAccessorType.Y1 ? `${label}${y1AccessorFormat}` : `${label}${y0AccessorFormat}`;
+  }
+
+  private renderLegendElement = (item: SeriesLegendItem) => {
+    const { key, displayValue, banded } = item;
+    const { legendPosition, legendItemTooltipValues, isCursorOnChart } = this.props.chartStore!;
+    const tooltipValues = legendItemTooltipValues.get();
+    const legendValues = this.getLegendValues(tooltipValues, key, banded);
+
+    return legendValues.map((value, index) => {
+      const yAccessor: BandedAccessorType = index === 0 ? BandedAccessorType.Y1 : BandedAccessorType.Y0;
+      return (
+        <LegendItem
+          {...item}
+          label={this.getItemLabel(item, yAccessor)}
+          key={`${key}-${yAccessor}`}
+          legendItemKey={key}
+          legendPosition={legendPosition.get()}
+          displayValue={isCursorOnChart.get() ? value : displayValue.formatted[yAccessor]}
+          onMouseEnter={this.onLegendItemMouseover(key)}
+          onMouseLeave={this.onLegendItemMouseout}
+        />
+      );
+    });
   };
 }