+++ stories/bar_chart.tsx
@@ -23,7 +23,6 @@ import {
   niceTimeFormatByDay,
   Position,
   RectAnnotation,
-  Rotation,
   ScaleType,
   Settings,
   timeFormatter,
@@ -35,6 +34,8 @@ import { KIBANA_METRICS } from '../src/utils/data_samples/test_dataset_kibana';
 
 import { TEST_DATASET_DISCOVER } from '../src/utils/data_samples/test_dataset_discover_per_30s';
 import { FilterPredicate } from '../src/chart_types/xy_chart/utils/specs';
+import { getRandomNumber } from '../src/utils/data_generators/simple_noise';
+import { getChartRotationKnob } from './common';
 
 const dateFormatter = timeFormatter('HH:mm:ss');
 
@@ -77,7 +78,7 @@ storiesOf('Bar Chart', module)
     );
   })
   .add('with value label', () => {
-    const showValueLabel = boolean('show value label', false);
+    const showValueLabel = boolean('show value label', true);
     const isAlternatingValueLabel = boolean('alternating value label', false);
     const isValueContainedInElement = boolean('contain value label within bar element', false);
     const hideClippedValue = boolean('hide clipped value', false);
@@ -89,17 +90,7 @@ storiesOf('Bar Chart', module)
       hideClippedValue,
     };
 
-    const debug = boolean('debug', true);
-    const chartRotation = select<Rotation>(
-      'chartRotation',
-      {
-        '0 deg': 0,
-        '90 deg': 90,
-        '-90 deg': -90,
-        '180 deg': 180,
-      },
-      0,
-    );
+    const debug = boolean('debug', false);
 
     const theme = {
       barSeriesStyle: {
@@ -134,7 +125,7 @@ storiesOf('Bar Chart', module)
 
     return (
       <Chart renderer="canvas" className={'story-chart'}>
-        <Settings theme={theme} debug={debug} rotation={chartRotation} />
+        <Settings theme={theme} debug={debug} rotation={getChartRotationKnob()} showLegend />
         <Axis id={getAxisId('bottom')} position={Position.Bottom} title={'Bottom axis'} showOverlappingTicks={true} />
         <Axis
           id={getAxisId('left2')}
@@ -226,17 +217,6 @@ storiesOf('Bar Chart', module)
     );
   })
   .add('with linear x axis', () => {
-    const chartRotation = select<Rotation>(
-      'chartRotation',
-      {
-        '0 deg': 0,
-        '90 deg': 90,
-        '-90 deg': -90,
-        '180 deg': 180,
-      },
-      0,
-    );
-
     const theme = {
       ...LIGHT_THEME,
       scales: {
@@ -256,7 +236,7 @@ storiesOf('Bar Chart', module)
     };
     return (
       <Chart className={'story-chart'}>
-        <Settings rotation={chartRotation} theme={theme} />
+        <Settings rotation={getChartRotationKnob()} theme={theme} />
         <Axis id={getAxisId('bottom')} position={Position.Bottom} title={'Bottom axis'} showOverlappingTicks={true} />
         <Axis
           id={getAxisId('left2')}
@@ -509,17 +489,6 @@ storiesOf('Bar Chart', module)
     );
   })
   .add('clustered with axis and legend', () => {
-    const chartRotation = select<Rotation>(
-      'chartRotation',
-      {
-        '0 deg': 0,
-        '90 deg': 90,
-        '-90 deg': -90,
-        '180 deg': 180,
-      },
-      0,
-    );
-
     const theme = {
       ...LIGHT_THEME,
       scales: {
@@ -540,7 +509,7 @@ storiesOf('Bar Chart', module)
 
     return (
       <Chart className={'story-chart'}>
-        <Settings showLegend={true} legendPosition={Position.Right} theme={theme} rotation={chartRotation} />
+        <Settings showLegend={true} legendPosition={Position.Right} theme={theme} rotation={getChartRotationKnob()} />
         <Axis id={getAxisId('bottom')} position={Position.Bottom} title={'Bottom axis'} showOverlappingTicks={true} />
         <Axis
           id={getAxisId('left2')}
@@ -934,17 +903,6 @@ storiesOf('Bar Chart', module)
         }
       : undefined;
 
-    const chartRotation = select<Rotation>(
-      'chartRotation',
-      {
-        '0 deg': 0,
-        '90 deg': 90,
-        '-90 deg': -90,
-        '180 deg': 180,
-      },
-      0,
-    );
-
     const theme = {
       scales: {
         barsPadding: number('bars padding', 0.25, {
@@ -958,7 +916,7 @@ storiesOf('Bar Chart', module)
 
     return (
       <Chart className={'story-chart'}>
-        <Settings xDomain={xDomain} rotation={chartRotation} theme={theme} />
+        <Settings xDomain={xDomain} rotation={getChartRotationKnob()} theme={theme} />
         <Axis id={getAxisId('bottom')} position={Position.Bottom} title={'Bottom axis'} />
         <Axis
           id={getAxisId('left2')}
@@ -983,17 +941,6 @@ storiesOf('Bar Chart', module)
     const hasCustomDomain = boolean('has custom domain', false);
     const xDomain = hasCustomDomain ? ['a', 'b'] : undefined;
 
-    const chartRotation = select<Rotation>(
-      'chartRotation',
-      {
-        '0 deg': 0,
-        '90 deg': 90,
-        '-90 deg': -90,
-        '180 deg': 180,
-      },
-      0,
-    );
-
     const theme = {
       scales: {
         barsPadding: number('bars padding', 0.25, {
@@ -1007,7 +954,7 @@ storiesOf('Bar Chart', module)
 
     return (
       <Chart className={'story-chart'}>
-        <Settings xDomain={xDomain} rotation={chartRotation} theme={theme} />
+        <Settings xDomain={xDomain} rotation={getChartRotationKnob()} theme={theme} />
         <Axis id={getAxisId('bottom')} position={Position.Bottom} title={'Bottom axis'} />
         <Axis
           id={getAxisId('left2')}
@@ -1156,16 +1103,8 @@ storiesOf('Bar Chart', module)
     );
   })
   .add('scale to extent', () => {
-    const yScaleToDataExtent = boolean('yScaleDataToExtent', false);
-    const mixed = [
-      { x: 3, y: 1 },
-      { x: 0, y: -4 },
-      { x: 2, y: 2 },
-      { x: 1, y: -3 },
-      { x: 2, y: 2 },
-      { x: 1, y: -3 },
-      { x: 3, y: 1 },
-    ];
+    const yScaleToDataExtent = boolean('yScaleDataToExtent', true);
+    const mixed = [{ x: 0, y: -4 }, { x: 1, y: -3 }, { x: 2, y: 2 }, { x: 3, y: 1 }];
 
     const allPositive = mixed.map((datum) => ({ x: datum.x, y: Math.abs(datum.y) }));
     const allNegative = mixed.map((datum) => ({ x: datum.x, y: Math.abs(datum.y) * -1 }));
@@ -1177,7 +1116,7 @@ storiesOf('Bar Chart', module)
         allPositive: 'all positive',
         allNegative: 'all negative',
       },
-      'mixed',
+      'all negative',
     );
 
     let data = mixed;
@@ -1218,8 +1157,8 @@ storiesOf('Bar Chart', module)
     const data = KIBANA_METRICS.metrics.kibana_os_load[0].data.map((d: any) => {
       return {
         x: d[0],
-        max: d[1] + 4 + 4 * Math.random(),
-        min: d[1] - 4 - 4 * Math.random(),
+        max: d[1] + 4 + 4 * getRandomNumber(),
+        min: d[1] - 4 - 4 * getRandomNumber(),
       };
     });
     const lineData = KIBANA_METRICS.metrics.kibana_os_load[0].data.map((d: any) => {
@@ -1482,17 +1421,6 @@ storiesOf('Bar Chart', module)
       },
     };
 
-    const chartRotation = select<Rotation>(
-      'chartRotation',
-      {
-        '0 deg': 0,
-        '90 deg': 90,
-        '-90 deg': -90,
-        '180 deg': 180,
-      },
-      0,
-    );
-
     const theme = {
       scales: {
         barsPadding: number('bars padding', 0.25, {
@@ -1549,7 +1477,7 @@ storiesOf('Bar Chart', module)
 
     return (
       <Chart className={'story-chart'}>
-        <Settings rotation={chartRotation} theme={theme} debug={boolean('debug', true)} />
+        <Settings rotation={getChartRotationKnob()} theme={theme} debug={boolean('debug', true)} />
         <LineAnnotation
           annotationId={getAnnotationId('line-annotation')}
           domainType={AnnotationDomainTypes.XDomain}
@@ -1613,18 +1541,6 @@ storiesOf('Bar Chart', module)
   })
   .add('[test] histogram mode (ordinal)', () => {
     const data = [{ x: 'a', y: 2 }, { x: 'b', y: 7 }, { x: 'c', y: 0 }, { x: 'd', y: 6 }];
-
-    const chartRotation = select<Rotation>(
-      'chartRotation',
-      {
-        '0 deg': 0,
-        '90 deg': 90,
-        '-90 deg': -90,
-        '180 deg': 180,
-      },
-      0,
-    );
-
     const theme = {
       scales: {
         barsPadding: number('bars padding', 0.25, {
@@ -1640,7 +1556,7 @@ storiesOf('Bar Chart', module)
 
     return (
       <Chart className={'story-chart'}>
-        <Settings rotation={chartRotation} theme={theme} debug={boolean('debug', true)} />
+        <Settings rotation={getChartRotationKnob()} theme={theme} debug={boolean('debug', true)} />
         <Axis id={getAxisId('discover-histogram-left-axis')} position={Position.Left} title={'left axis'} />
         <Axis id={getAxisId('discover-histogram-bottom-axis')} position={Position.Bottom} title={'bottom axis'} />
         {hasHistogramBarSeries && (